fix: workers creation race (#80)
Consider the following scenario: we have two tests that should run in
different workers.

Currently, the following race is possible:
1. We create a worker for the first test and run it.
2. We need a worker for the second test, and since first
   worker is busy, we spawn a new one.
3. While second worker is initializing, the first one
   finishes and is ready for reuse.
4. We grab the first worker for the second test. However,
   it does not match the hash, so we spawn the third worker.

Overall, we get three workers instead of two. The solution
is to specifically wait for the newly created worker, and
leave reusable workers for future tests.

This race was often seen on the bots when running
"should work with two different fixture objects".