feat: reuse worker process between test types
When two test types only differ in the last few
environments, and none of them has `beforeAll` or `afterAll`,
we can reuse the worker. In this case, the worker won't have
to run new `beforeAll`/`afterAll` methods, and will update
the list of envs to run `beforeEach`/`afterEach` for the next test.